Drop the dependency on FreeBSD's pkg command line tool.
Now that OS X doesn't ship with OpenSSL anymore, I don't want to pull in
even more dependencies. Let's just craft the packages manually, like we
do for the other systems. The packages are identical to the ones created
by pkg, with the exception that the dependencies are now stored in
sorted order.