Remove host system binaries from PATH.
In order to get better reproducible results, let's not use any of the
host system binaries (with small exceptions) to build packages. Adjust
PATH to just point to /usr/obj/cloudabi-ports/bin. Make sure we depend
on a whole bunch of GNU packages, so that all necessary tools are
available.