Properly generate the list of URLs for a distfile.
For the fallback site, we do need to make sure that we append the entire
distfile name -- not just the basename. I just ran into this issue,
because libtom.net is down. All of these distfiles have been prefixed.
The build script was unable to fetch them from nuxi.nl.