Don't let the configure test for GMP depend on a nonexistent symbol.
It looks like this specific autoconf test depends on the existence of
the open() function. The code is not referenced, so this test would
normally pass because GNU ld doesn't care about it if we use
--gc-sections.

LLD 3.8, on the other hand, is a bit more picky. It still reports
missing symbols in unreferenced code, even if --gc-sections is provided.