Already import LLVM 3.8.0rc1 under the version 3.7.91.
LLVM 3.8 is going to be pretty important for us:

- We can discard all of our local compiler patches.
- We can start using LLD at least on x86-64.

Let's go ahead and import the first release candidate of LLVM 3.8. The
final version is expected to come out next month. Also install LLD. LLD
is compatible with GNU ld, but requires us to pass in -flavor gnu. Add a
small wrapper application, ld.c, that forwards calls to the linker to
LLD with -flavor gnu set.