Use Clang 3.7 as the host C/C++ compiler on Linux-based systems.
GCC has the disadvantage that as(1) and ld(1) need to remain in $PATH in
order for it to work. This is problematic in our case, as we try to
explicitly shield off utilities that are not part of the build
infrastructure.

Switch over to using Clang, which does not have this dependency. This
fixes the build failure reported by Alex Willmer in issue #5.