Clean up the buildroot directory in case we finish successfully.
Alex Willmer reported in issue #9 that build_packages.py leaves files
behind on the filesystem when finished. This was actually done on
purpose, as we don't want to remove the build directory in case the
build failed.

That said, if build_packages.py do finishes successfully, the buildroot
directory rarely contains anything useful; merely the data files that
were used to craft the new package. In that case it makes a lot of sense
to just remove the directory.