Drop the dependency on the cloudlibc package.
In the upcoming version of cloudlibc I want to be able to install the
unit test binary as well, so we can perform operating system testing
against released versions of cloudlibc. Because of that, we need to make
sure that we can build a copy of compiler-rt without depending on
anything.