Append package revision numbers unconditionally.
For systems like Debian and Arch Linux it is uncommon to start numbering
package revisions at 0. They use 1 instead. We worked around this on
Debian by hiding the revision number if it was 0, which wasn't quite
right.

Change the code to start numbering package revisions at 1 and append it
to the package names unconditionally. Once we've performed a full
package build with this change, we can consider simplifying version
number parsing as well.