Add a proof of concept port of memcached.
Memcached is a rather simple tool that does not have any dependencies on
filesystem access. It also doesn't craft any sockets, except on startup.
This makes it an ideal application to be sandboxed.

This patched up version attempts to integrate it with argdata.h, but
fails miserably at it. I still have to convert the getopt() loop that is
in main() to be based on argdata.h. Right now we assume file descriptor
0 is the socket, which is also pretty bad.