Properly parse the argument data provided to memcached.
I think we now have a version of memcached that is complete enough to
act as a prototype. It seems to work for me, but that said, I've never
used memcached before. I am not a power user.

Let's keep this port around until someone who uses memcached fanatically
feels like improving this port. In the meantime, we could already start
upstreaming some of the smaller changes.