Use Position Independent Executables exclusively.
Import a newer version of LLVM that has a better version of LLD. This
version of LLD does PIE on both x86-64 and aarch64 flawlessly, with the
exception of a tiny flaw in how it emits relocations. Patch this up
locally for now.

Once this is fixed upstream, I'll also commit the other change to enable
PIE in the Clang frontend unconditionally. That way we can finally
switch back to a stock copy of LLVM.

This also means that we're switching to LLD on aarch64.