Enable the '_crypt' module.
As we now have a C library that provides crypt_r(), let's patch up
Python to make use of it. The existing '_crypt' module that provides the
native binding for crypt.crypt() still uses the thread-unsafe C library
function, so patch it up to make use of crypt_r().

We also need to patch up this module to properly include <unistd.h>. It
normally gets away with an implicit prototype, but because it needs to
make use of a 'struct crypt_data' now, it now needs the header
explicitly.

Still disable the call to crypt() when doing the bootstrap build, as on
most systems, crypt() is provided by an external libcrypt. This
contradicts POSIX, which requires that crypt() is part of the C library.