Add support for propagating argument data into Python.
At the root, add a key called "args". This may contain a subtree of
arguments that are passed on to the Python script under 'sys.argdata'.

Ensure that these file descriptors are created with closefd=0, as they
may be referenced more than once. Also disable buffering for now, as
buffering only works in combination with seekable file descriptors.

Create socket.socket objects for socket file descriptors.