Don't build Python for the host system with our Setup.dist.
This file should only be respected when doing the actual target build.
Otherwise we start depending on a whole bunch of libraries on the host
system, just to be able to build for the target.

While there, properly make use of our packaged version of Expat instead
of the outdated version shipped with Python itself.