Update LLVM-provided libraries to 3.9.0rc1.
For libcxx, we can now drop a single change that has been upstreamed.
Unfortunately, the addition of some new *.h wrapper causes some troubles
for us. We can now no longer include <string.h> from within C++ code
twice. This is problematic for us, as this is needed for the unsafe
string functions.