Upgrade the Python package to 3.6.0b1.
It looks like this release allows us to eliminate some of our local
changes. The datetime module now uses gmtime_r()/localtime_r() out of
the box. The changes regarding _freeze_importlib are now fully
integrated.

What is kind of weird is that we now need to build with signalmodule
enabled, as some functions got moved around in the Python source tree.
I'd like to see us get rid of that module entirely.