Don't forget to link in the accelerated CRC32 implementation.
Right now we get linker errors when using LevelDB, because it cannot
find the CRC32 implementation that needs to be linked into the library.
This can be fixed by also building port_posix_sse.c.

At the same time, there is a flaw in the SSE4.2 detection logic in
build_detect_platform. Clang doesn't complain when passing in -msse4.2
on ARM-based systems. Add an explicit preprocessor test to make this a
bit more strict.