Add os.pdwait() to Python.
pdwait() is a CloudABI-specific function to get the exit status of a child
process by process file descriptor. It is to pdfork() what waidpid() is to
fork(). The Python version of pdwait() takes a file descriptor and an options
parameter.

It returns None if the option WNOHANG is given and the child has not exited
yet, or a pdwait_result if it has. The pdwait_result is a Python structure
analogous to siginfo_t, which can be read as a tuple (si_signo, si_status,
si_code) or via the attributes names. The order of the tuple is the same as the
order in the waitid() tuple, to make it easier to write code compatible with
both.