Upgrade CloudABI components to the latest upstream versions.
The most important change is the removal of <aio.h> and kqueue() from
cloudlibc. This is in preparation for the integration of a libuv
compatible event loop into the C library, making these interfaces
superfluous.

In the process, remove a patch from libevent that has now become
unnecessary. The source file to which it applies is no longer built.